+++ settingrepo-impl/src/main/java/com/dwarfeng/settingrepo/impl/handler/TextNodeOperateHandlerImpl.java
@@ -11,6 +11,7 @@ import com.dwarfeng.settingrepo.stack.handler.TextNodeOperateHandler;
 import com.dwarfeng.settingrepo.stack.service.SettingNodeMaintainService;
 import com.dwarfeng.settingrepo.stack.service.TextNodeMaintainService;
 import com.dwarfeng.subgrade.sdk.exception.HandlerExceptionHelper;
+import com.dwarfeng.subgrade.sdk.interceptor.analyse.BehaviorAnalyse;
 import com.dwarfeng.subgrade.stack.bean.key.StringIdKey;
 import com.dwarfeng.subgrade.stack.exception.HandlerException;
 import org.springframework.stereotype.Component;
@@ -47,6 +48,7 @@ public class TextNodeOperateHandlerImpl implements TextNodeOperateHandler {
     }
 
     @SuppressWarnings("DuplicatedCode")
+    @BehaviorAnalyse
     @Override
     public TextNodeInspectResult inspect(TextNodeInspectInfo info) throws HandlerException {
         try {
@@ -85,6 +87,7 @@ public class TextNodeOperateHandlerImpl implements TextNodeOperateHandler {
         }
     }
 
+    @BehaviorAnalyse
     @Override
     public void put(TextNodePutInfo info) throws HandlerException {
         try {